# playbook-studio/studio-frontend/src/models/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "playbook-studio API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/skills": {
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "Create Skill",
        "description": "Run the bootstrap-skill transaction; gated like every edit.",
        "operationId": "create_skill_api_skills_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSkillRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/skills/{name}": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "Get Skill",
        "description": "Return one parsed skill by name.",
        "operationId": "get_skill_api_skills__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "skills"
        ],
        "summary": "Update Skill Document",
        "description": "Replace the whole SKILL.md; the edit survives only if the gates pass.",
        "operationId": "update_skill_document_api_skills__name__put",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSkillRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Refs"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "refs"
        ],
        "title": "CreateSkillRequest",
        "description": "Inputs for the bootstrap-skill transaction."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SkillResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Refs"
          },
          "raw": {
            "type": "string",
            "title": "Raw"
          },
          "body": {
            "type": "string",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "refs",
          "raw",
          "body"
        ],
        "title": "SkillResponse",
        "description": "A parsed SKILL.md returned over HTTP."
      },
      "UpdateDocumentRequest": {
        "properties": {
          "raw": {
            "type": "string",
            "title": "Raw"
          }
        },
        "type": "object",
        "required": [
          "raw"
        ],
        "title": "UpdateDocumentRequest",
        "description": "The new full markdown content (frontmatter included) for one REF."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationReportResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "tests_passed": {
            "type": "boolean",
            "title": "Tests Passed"
          },
          "links_ok": {
            "type": "boolean",
            "title": "Links Ok"
          },
          "tests_output": {
            "type": "string",
            "title": "Tests Output"
          },
          "links_output": {
            "type": "string",
            "title": "Links Output"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "tests_passed",
          "links_ok",
          "tests_output",
          "links_output"
        ],
        "title": "ValidationReportResponse",
        "description": "Outcome of the playbook's validation gates returned over HTTP."
      }
    }
  }
}